+++ packages/xbell-reporter/lib/const.js
@@ -1,0 +1,5 @@
+const XBELL_RESOURCE_PLACEHOLDER = '\<\!-- xbell-resources-placeholder --\>';
+
+module.exports = {
+  XBELL_RESOURCE_PLACEHOLDER
+}