+++ packages/xbell-reporter/src/components/SearchAppBar.tsx
@@ -1,0 +1,119 @@
+import React from 'react';
+import { styled, alpha } from '@mui/material/styles';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import IconButton from '@mui/material/IconButton';
+import Typography from '@mui/material/Typography';
+import InputBase from '@mui/material/InputBase';
+import Tabs from '@mui/material/Tabs';
+import Tab from '@mui/material/Tab';
+import NotificationsActiveIcon from '@mui/icons-material/NotificationsActive';
+// import SearchIcon from '@mui/icons-material/Search';
+import ThemeSwitch from './ThemeSwitch';
+
+const Search = styled('div')(({ theme }) => ({
+  position: 'relative',
+  borderRadius: theme.shape.borderRadius,
+  backgroundColor: alpha(theme.palette.common.white, 0.15),
+  '&:hover': {
+    backgroundColor: alpha(theme.palette.common.white, 0.25),
+  },
+  marginLeft: 0,
+  width: '100%',
+  [theme.breakpoints.up('sm')]: {
+    marginLeft: theme.spacing(1),
+    width: 'auto',
+  },
+}));
+
+const SearchIconWrapper = styled('div')(({ theme }) => ({
+  padding: theme.spacing(0, 2),
+  height: '100%',
+  position: 'absolute',
+  pointerEvents: 'none',
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+}));
+
+const StyledInputBase = styled(InputBase)(({ theme }) => ({
+  color: 'inherit',
+  '& .MuiInputBase-input': {
+    padding: theme.spacing(1, 1, 1, 0),
+    // vertical padding + font size from searchIcon
+    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
+    transition: theme.transitions.create('width'),
+    width: '100%',
+    [theme.breakpoints.up('sm')]: {
+      width: '12ch',
+      '&:focus': {
+        width: '20ch',
+      },
+    },
+  },
+}));
+
+const StyledTabs = styled(Tabs)(({ theme }) => ({
+    width: '400px',
+}));
+
+interface SearchAppBarProps {
+  envs: string[];
+  activeEnv: string;
+  onEnvChange(env: string): void;
+}
+
+export default function SearchAppBar({
+  envs,
+  activeEnv,
+  onEnvChange,
+}: SearchAppBarProps) {
+  return (
+    <Box sx={{ flexGrow: 1 }}>
+      <AppBar position="static">
+        <Toolbar>
+          <IconButton
+            size="large"
+            edge="start"
+            aria-label="open drawer"
+            sx={{ mr: 2 }}
+          >
+            <NotificationsActiveIcon />
+          </IconButton>
+          <Typography
+            variant="h6"
+            noWrap
+            component="div"
+            sx={{ display: { xs: 'none', sm: 'block' } }}
+          >
+            X - BELL
+          </Typography>
+          <Box style={{ flexGrow: 1, display: 'flex', justifyContent: 'center' }}>
+            <StyledTabs
+              centered
+              value={envs.indexOf(activeEnv)}
+              onChange={(_, v) => onEnvChange(envs[v])}
+              indicatorColor="secondary"
+              textColor="inherit"
+            >
+              {envs.map((env) => (
+                <Tab key={env} label={env.toLocaleUpperCase()} />
+              ))}
+            </StyledTabs>
+          </Box>
+          {/* <ThemeSwitch /> */}
+          {/* <Search>
+            <SearchIconWrapper>
+              <SearchIcon />
+            </SearchIconWrapper>
+            <StyledInputBase
+              placeholder="搜索"
+              inputProps={{ 'aria-label': 'search' }}
+            />
+          </Search> */}
+        </Toolbar>
+      </AppBar>
+    </Box>
+  );
+}