+++ packages/xbell/src/core/reporter.ts
@@ -1,4 +1,3 @@
-import { } from '../utils/print';
 import * as chalk from 'chalk';
 
 export enum Status {