+++ packages/xbell/CHANGELOG.md
@@ -1,5 +1,16 @@
 # xbell
 
+## 0.2.0
+
+### Minor Changes
+
+- b774699: Generate test reports that can watch video record
+
+### Patch Changes
+
+- Updated dependencies [b774699]
+  - xbell-reporter@0.1.1
+
 ## 0.1.0
 
 ### Patch Changes